requirements.txt
toml>=0.7
Mail>=2.1.0
flask>=0.11.1
flask-mail>=0.9.1
flask_sqlalchemy>=2.1
flask_user==0.6.8
flask-login==0.3.2
flask-wtf>=0.12
flask-pytest>=0.0.5
pytest-flask>=0.10.0
Flask-Babel>=0.11.1
gitpython>=2.0.8
sphinxcontrib-tikz>=0.4.2
mysql-python
cloud-sptheme>=1.7.1
docutils-tinyhtmlwriter>=1.2.0
Flask-Limiter>=0.9.3
Flask-Script==2.0.5
arrow>=0.10.0
creole>=1.2
python-creole>=1.3.1
Flask-Assets==0.12
cssmin==0.2.0
closure==20161201
dulwich==0.18.2

